CSN: CSN-m5-74
Modified the m5 global event queue to not use the __thread attribute.
This attribute creates a new instance of the event queue pointer for
each thread that touches the event queue.  This was breaking new-style
LIM compilations, since the new compilations may choose to multithread
the old RRR service handlers.